Owed money amount is now float instead of int

+++ src/Entity/Modules/Payments/MyPaymentsOwed.php
@@ -37,7 +37,7 @@ class MyPaymentsOwed
     private $date;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $amount;
 
@@ -105,12 +105,12 @@ class MyPaymentsOwed
         return $this;
     }
 
-    public function getAmount(): ?int
+    public function getAmount(): ?float
     {
         return $this->amount;
     }
 
-    public function setAmount(int $amount): self
+    public function setAmount(float $amount): self
     {
         $this->amount = $amount;
